Added comments tests

+++ profile-matcher-core/src/main/java/com/biit/profile/core/controllers/ProfileCandidateCommentController.java
@@ -3,8 +3,11 @@ package com.biit.profile.core.controllers;
 
 import com.biit.profile.core.converters.ProfileCandidateCommentConverter;
 import com.biit.profile.core.converters.models.ProfileCandidateCommentConverterRequest;
+import com.biit.profile.core.exceptions.CandidateNotFoundException;
+import com.biit.profile.core.exceptions.CommentTooLongException;
 import com.biit.profile.core.models.ProfileCandidateCommentDTO;
 import com.biit.profile.core.providers.ProfileCandidateCommentProvider;
+import com.biit.profile.core.providers.ProfileCandidateProvider;
 import com.biit.profile.persistence.entities.ProfileCandidateComment;
 import com.biit.profile.persistence.entities.ProfileCandidateId;
 import com.biit.profile.persistence.repositories.ProfileCandidateCommentRepository;
@@ -17,8 +20,12 @@ import java.util.Set;
 public class ProfileCandidateCommentController extends ElementController<ProfileCandidateComment, ProfileCandidateId, ProfileCandidateCommentDTO,
         ProfileCandidateCommentRepository, ProfileCandidateCommentProvider, ProfileCandidateCommentConverterRequest, ProfileCandidateCommentConverter> {
 
-    protected ProfileCandidateCommentController(ProfileCandidateCommentProvider provider, ProfileCandidateCommentConverter converter) {
+    private final ProfileCandidateProvider profileCandidateProvider;
+
+    protected ProfileCandidateCommentController(ProfileCandidateCommentProvider provider, ProfileCandidateCommentConverter converter,
+                                                ProfileCandidateProvider profileCandidateProvider) {
         super(provider, converter);
+        this.profileCandidateProvider = profileCandidateProvider;
     }
 
     @Override
@@ -40,8 +47,17 @@ public class ProfileCandidateCommentController extends ElementController<Profile
 
 
     public ProfileCandidateComment addComment(Long profileId, Long userId, String comment) {
+        if (comment.length() > ProfileCandidateComment.COMMENT_LENGTH) {
+            throw new CommentTooLongException(this.getClass(), "Comment length exceeds the limit of '"
+                    + ProfileCandidateComment.COMMENT_LENGTH + "' bytes");
+        }
+
         //Delete previous one if exists.
         getProvider().deleteByIdProfileIdAndIdUserId(profileId, userId);
+
+        //Checks that exist the candidate
+        profileCandidateProvider.findByProfileIdAndUserId(profileId, userId).orElseThrow(() ->
+                new CandidateNotFoundException(this.getClass(), "No candidate '" + userId + "' found for profile '" + profileId + "'."));
 
         return getProvider().save(new ProfileCandidateComment(profileId, userId, comment));
     }
